used offers from store in FavoritesScreen; renamed TSortTypes -> SortTypesType and useOffersSort -> sortOffers
