+++ src/components/molecules/offers-list-with-sort/offers-list-with-sort.tsx
@@ -1,8 +1,8 @@
 import { OffersList, SortSelector } from '..';
 import { OfferData } from '../../../mocks';
-import {CitiesType, SORT_TYPES, TSortTypes} from '../../../constants';
-import {useState} from 'react';
-import {useOffersSort} from '../../../hooks/useOffersSort.ts';
+import { CitiesType, SORT_TYPES, SortTypesType } from '../../../constants';
+import { useState } from 'react';
+import { sortOffers } from '../../../utils';
 
 type OfferCardsListProps = {
   offers: OfferData[];
@@ -13,9 +13,9 @@ type OfferCardsListProps = {
 
 function OffersListWithSort({offers, city, handleCardHover, handleCardLeave}: OfferCardsListProps) {
 
-  const [currentSort, setCurrentSort] = useState<TSortTypes>(SORT_TYPES.Popular);
+  const [currentSort , setCurrentSort] = useState<SortTypesType>(SORT_TYPES.Popular);
 
-  const handleSetSort = (sort: TSortTypes) => {
+  const handleSetSort = (sort: SortTypesType) => {
     setCurrentSort(sort);
   };
 
@@ -25,7 +25,7 @@ function OffersListWithSort({offers, city, handleCardHover, handleCardLeave}: Of
       <b className="places__found">{offers.length} places to stay in {city}</b>
       <SortSelector selectedSort={currentSort} onSelect={handleSetSort}/>
       <OffersList
-        offers={useOffersSort(currentSort, offers)}
+        offers={sortOffers(currentSort, offers)}
         classNames={'cities__places-list places__list tabs__content'}
         handleHoverOnCard={handleCardHover}
         handleCardLeave={handleCardLeave}
