+++ src/constants/index.ts
@@ -5,7 +5,7 @@ import AUTH_STATUS from './auth-status.ts';
 import { URL_MARKER_CURRENT } from './map-icons.ts';
 import { URL_MARKER_DEFAULT } from './map-icons.ts';
 import { CitiesType } from './cities.ts';
-import { TSortTypes } from './sort.ts';
+import { SortTypesType } from './sort.ts';
 
 export {
   CITIES,
@@ -18,5 +18,5 @@ export {
 
 export type {
   CitiesType,
-  TSortTypes,
+  SortTypesType,
 };
