+++ src/utils/index.ts
@@ -1,7 +1,9 @@
-import ScrollToTop from './use-scroll-to-top.ts';
+import ScrollToTop from './scroll-to-top.ts';
 import { capitalize } from './string.ts';
+import { sortOffers } from './sort-offers.ts';
 
 export {
   ScrollToTop,
   capitalize,
+  sortOffers,
 };
